requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
boto3>=1.28.0
python-dotenv>=1.0.0
mangum>=0.17.0
requests>=2.31.0
httpx>=0.25.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
moto>=4.2.0
httpx>=0.25.0
faker>=19.0.0
pytest-xdist>=3.0.0
pytest-timeout>=2.1.0

